+++ examples/body.rs
@@ -1,4 +1,5 @@
 use foxhole::{
+    connection::Http1,
     resolve::{Post, Resolve, ResolveGuard},
     run, sys, PathIter, RequestState, Route,
 };
@@ -11,7 +12,7 @@ impl<'a, 'b> Resolve<'b> for Body<'a> {
     type Output = Body<'b>;
 
     fn resolve(ctx: &'b RequestState, _path_iter: &mut PathIter) -> ResolveGuard<Self::Output> {
-        let body = str::from_utf8(ctx.request.body().get().as_slice()).unwrap();
+        let body = str::from_utf8(ctx.request.body().get_as_slice()).unwrap();
 
         ResolveGuard::Value(Body(body))
     }
@@ -26,5 +27,5 @@ fn post(_post: Post, body: Body) -> u16 {
 fn main() {
     let route = Route::new(sys![post]);
 
-    run("127.0.0.1:8080", route);
+    run::<Http1>("127.0.0.1:8080", route);
 }