+++ src/lib.rs
@@ -8,10 +8,13 @@ pub mod resolve;
 pub mod routing;
 pub mod systems;
 pub mod type_cache;
+pub mod connection;
+pub mod get_as_slice;
 
 mod lazy;
 mod sequential_writer;
 mod tasks;
+mod tls_connection;
 
 pub use tasks::PathIter;